planning: fix compile warning.

+++ modules/planning/navi_planning.cc
@@ -208,7 +208,7 @@ void NaviPlanning::RunOnce(const LocalView& local_view,
       vehicle_state, start_timestamp, planning_cycle_time,
       last_publishable_trajectory_.get(), &replan_reason);
 
-  const uint32_t frame_num = seq_num_++;
+  const uint32_t frame_num = static_cast<uint32_t>(seq_num_++);
   status = InitFrame(frame_num, stitching_trajectory.back(), start_timestamp,
                      vehicle_state, trajectory_pb);
   if (!frame_) {
